requirements.txt
#pip install -r requirements.txt
#

MinimalModbus>=0.7
pyserial>=3.4
pygobject>=3.20.0
lark-parser>=0.5.6
